fix 0

+++ components/tidb_query_aggr/src/impl_max_min.rs
@@ -514,8 +514,8 @@ where
                     self.extremum = value.copied()
                 }
             } else {
-                let v1 = self.extremum.map(|x| x as i64);
-                let v2 = value.map(|x| *x as i64);
+                let v1 = self.extremum.map(|x| x);
+                let v2 = value.map(|x| *x);
                 if v1.cmp(&v2) == E::ORD {
                     self.extremum = value.copied()
                 }
